+++ CHANGELOG.md
@@ -25,3 +25,5 @@ equivalent to 0.2.0
 
 - Added: `clang-tidy.buildPath` option to tell clang-tidy where the build
 directory is. This is equivalent to `clang-tidy -p ...`
+
+- Fix: Can't save while clang-tidy is running on large files