+++ CHANGELOG.md
@@ -29,3 +29,6 @@ equivalent to 0.2.0
 - Initial release
 
 ## [Unreleased]
+
+- Fixed: Clang-Tidy will only run once. Subsequent runs produce
+  `Command failed: taskkill /pid 19412 /f /t ERROR: The process "19412" not found.`